+++ src/main/java/org/realityforge/replicant/server/ee/Replicate.java
@@ -1,0 +1,19 @@
+package org.realityforge.replicant.server.ee;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import javax.interceptor.InterceptorBinding;
+
+/**
+ * Interceptor binding for replication interceptor.
+ */
+@Inherited
+@Target( { ElementType.TYPE, ElementType.METHOD } )
+@Retention( RetentionPolicy.RUNTIME )
+@InterceptorBinding
+public @interface Replicate
+{
+}